Move config-focused view specs into config/ dir
