+++ spec/integration/view/config/part_class_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe "App view / Part class", :app_integration do
+RSpec.describe "App view / Config / Part class", :app_integration do
   before do
     with_directory(make_tmp_directory) do
       write "config/app.rb", <<~RUBY
