+++ spec/integration/view/config/scope_class_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe "App view / Scope class", :app_integration do
+RSpec.describe "App view / Config / Scope class", :app_integration do
   before do
     with_directory(@dir = make_tmp_directory) do
       write "config/app.rb", <<~RUBY
